refactor(member, s3): 패키지 네이밍 infra -> infrastructure로 수정
