+++ src/main/java/com/lhgpds/algometa/configuration/S3Configuration.java
@@ -4,8 +4,8 @@ import com.amazonaws.auth.AWSStaticCredentialsProvider;
 import com.amazonaws.auth.BasicAWSCredentials;
 import com.amazonaws.services.s3.AmazonS3Client;
 import com.amazonaws.services.s3.AmazonS3ClientBuilder;
-import com.lhgpds.algometa.infra.s3.S3Properties;
-import com.lhgpds.algometa.infra.s3.S3Uploader;
+import com.lhgpds.algometa.infrastructure.s3.S3Properties;
+import com.lhgpds.algometa.infrastructure.s3.S3Uploader;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
